+++ src/frontend/apps/impress/src/features/docs/doc-editor/types.tsx
@@ -1,3 +1,7 @@
+import { BlockNoteEditor } from '@blocknote/core';
+
+import { blockNoteSchema } from './components/BlockNoteEditor';
+
 export interface DocAttachment {
   file: string;
 }
@@ -12,3 +16,9 @@ export type HeadingBlock = {
     level: number;
   };
 };
+
+export type DocsBlockNoteEditor = BlockNoteEditor<
+  typeof blockNoteSchema.blockSchema,
+  typeof blockNoteSchema.inlineContentSchema,
+  typeof blockNoteSchema.styleSchema
+>;